Replace hspec test suite with tasty-based test suite
